trippy lights

+++ framework/include/Color.hpp
@@ -1,6 +1,6 @@
 #ifndef OPENGL_FRAMEWORK_COLOR_HPP
 #define OPENGL_FRAMEWORK_COLOR_HPP
 struct Color{
-    unsigned int r, g, b;
+    float r, g, b;
 };
 #endif //OPENGL_FRAMEWORK_COLOR_HPP
